fix(config): segurança Sanctum e qualidade das configurações
- Sanctum: expiração de 525600 (1 ano) para 43200 min (30 dias)
- Sanctum: prefixo mvapi_ para detecção de vazamento via GitHub secret scanning
- Logging: cast int em LOG_DAILY_DAYS para type safety
- Horizon: documentar requisito de SESSION_DRIVER != array para auth via browser
- .env.example: adicionar variáveis SANCTUM_TOKEN_EXPIRATION e SANCTUM_TOKEN_PREFIX

+++ config/logging.php
@@ -32,7 +32,7 @@ return [
             'driver' => 'daily',
             'path' => storage_path('logs/laravel.log'),
             'level' => env('LOG_LEVEL', 'debug'),
-            'days' => env('LOG_DAILY_DAYS', 14),
+            'days' => (int) env('LOG_DAILY_DAYS', 14),
             'replace_placeholders' => true,
         ],
